+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `pulse-db-size` will be documented in this file
 
+## 1.0.2 - 2024-12-12
+
+- Fix query when ignored tables are missing
+- Minor refactorings
+
 ## 1.0.1 - 2024-12-12
 
 - Add oracle support